Keep route exports valid and scope launch docs CSS
